requirements.txt
asgiref==3.7.2
Django==4.2.7
django-ckeditor==6.7.0
django-js-asset==2.2.0
Pillow==10.1.0
sqlparse==0.4.4
typing-extensions==4.8.0
tzdata==2023.3

# For production
gunicorn
psycopg2-binary
whitenoise==6.2.0
Brotli==1.0.9
whitenoise[brotli]